src/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
asn1==3.2.0
attrs==26.1.0
certifi==2026.2.25
cffi==2.0.0 ; platform_python_implementation != 'PyPy'
charset-normalizer==3.4.7
contourpy==1.3.3
cryptography==46.0.7
cycler==0.12.1
enum-compat==0.0.3
fonttools==4.62.1
frozenlist==1.8.0
idna==3.11
kiwisolver==1.5.0
matplotlib==3.10.8
multidict==6.7.1
numpy==2.4.4
packaging==26.1
pillow==12.2.0
propcache==0.4.1
pycparser==3.0 ; implementation_name != 'PyPy' and platform_python_implementation != 'PyPy'
pyopenssl==26.0.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
reportlab==4.4.10
six==1.17.0
smart-meter-texas==0.5.5
tenacity==9.1.4
typing-extensions==4.15.0 ; python_full_version < '3.13'
yarl==1.23.0
